FIRST VERSION WORKING!!!!!!!!!!!! Added building and fixed a lot of bugs with it

+++ Oberien/src/model/Model.java
@@ -274,7 +274,7 @@ public class Model {
 			i = timeToBuild;
 		}
 		timeToBuild -= i;
-		life += maxLife/costMoney*i;
+		life += (float)maxLife/(float)costMoney*i;
 	}
 	
 	public void setDirection(int direction) {
